fix: Use --without instead of --only
If you use PEP621, there is no longer a `main` dependency group with that name, which breaks `poetry install --only main`. The alternative is to actively exclude the test and dev dependency groups, which is not optimal but fixes things.

This flow broke on my machine but not in CI. It's worth investigating why that is the case.