feat: transition to uv and copier (#271)
* feat: Move to UV + Copier

* chore: Reset package version

* fix: Run `starship init` in onCreateCommand

This ensures that by the time the Dev Container is created, the shell
prompt is effectively configured. If done in postCreateCommand, the
prompt is only configured after the container's creation and after `uv
sync` and `pre-commit install` are complete.

* fix: Remove `--offline` flag from `uv lock --check`

If the UV cache is not populated, the `uv lock --check --offline`
command will fail as it doesn't just check consistency, it also checks
the cache for the package metadata and fails on a cache miss due to the
`--offline` flag. Removing `--offline` simply means that in the case of
a cache miss, which will in principle be a rare occurrence, we will
fetch the metadata from the package repository,

```
× No solution found when resolving dependencies:
  ╰─▶ Because commitizen was not found in the cache and your project depends
      on commitizen>=4.1.0, we can conclude that your project's requirements
      are unsatisfiable.

      hint: Packages were unavailable because the network was disabled. When
      the network is disabled, registry packages may only be read from the
      cache."
```

* docs: add migration section

* docs: rename project

* test: reintroduce cookiecutter template testing

* docs: fix minor issues

---------

Co-authored-by: Xavier Goás Aguililla <xavier@radix.ai>
Co-authored-by: Laurent Sorber <laurent.sorber@gmail.com>